fix: 고양이 삭제하기 로직에 적절한 유저인지 검증하는 기능 추가
deleteCat 로직에 요청을 보낸 유저가 적절한 유저인지 검증하는 부분이 빠져있어 추가

+++ server/catvillage/src/main/java/com/twentyfour_seven/catvillage/cat/service/CatService.java
@@ -48,7 +48,14 @@ public class CatService {
         return findCat;
     }
 
-    public void removeCat(long catId) {
+    public void removeCat(long catId, String email) {
+        // 요청을 보낸 유저가 적절한지 검증
+        userService.findVerifiedEmail(email);
+        String userEmail = findVerifiedCat(catId).getUser().getEmail();
+        if (email.equals(userEmail)) {
+            throw new BusinessLogicException(ExceptionCode.INVALID_USER);
+        }
+
         Cat findCat = findVerifiedCat(catId);
         catRepository.delete(findCat);
     }
